+++ TadukooLookAndFeel/pom.xml
@@ -14,6 +14,10 @@
 			<groupId>${project.groupId}</groupId>
 			<artifactId>TadukooLang</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>TadukooView</artifactId>
+		</dependency>
 	</dependencies>
 	<build>
 		<sourceDirectory>${directory.source}</sourceDirectory>
