+++ TadukooLookAndFeel/src/com/github/tadukoo/util/lookandfeel/paintui/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains {@link com.github.tadukoo.util.lookandfeel.paintui.PaintUIResource PaintUIResource} and its implementations
+ * to allow for customizing the {@link java.awt.Paint Paint}s to be used in
+ * {@link com.github.tadukoo.util.lookandfeel.TadukooLookAndFeel TadukooLookAndFeel} within the
+ * {@link com.github.tadukoo.util.lookandfeel.TadukooTheme TadukooTheme}.
+ */
+package com.github.tadukoo.util.lookandfeel.paintui;
