+++ pom.xml
@@ -92,6 +92,12 @@
 				<artifactId>TadukooUtil</artifactId>
 				<version>${project.version}</version>
 			</dependency>
+			<!-- Tadukoo View -->
+			<dependency>
+				<groupId>${project.groupId}</groupId>
+				<artifactId>TadukooView</artifactId>
+				<version>${project.version}</version>
+			</dependency>
 			<!-- JUnit Dependencies -->
 			<dependency>
 				<groupId>org.junit.jupiter</groupId>
